안쓰는 옵션 제거

+++ Assets/Scripts/IO/FileParser.cs
@@ -63,7 +63,7 @@ public class FileParser : FileReader
 
                 if ( Contains( "MinBPM:" ) ) _song.minBpm    = int.Parse( SplitAndTrim( ':' ) );
                 if ( Contains( "MaxBPM:" ) ) _song.maxBpm    = int.Parse( SplitAndTrim( ':' ) );
-                if ( Contains( "Median:" ) ) _song.medianBpm = int.Parse( SplitAndTrim( ':' ) );
+                if ( Contains( "Median:" ) ) _song.medianBpm = double.Parse( SplitAndTrim( ':' ) );
             }
     }
         catch ( Exception _error )
